fetch/metadata/generated/element-iframe.https.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/element-iframe.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title></title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, test, userActivated) {
    const iframe = document.createElement('iframe');

    const setSrc = () => iframe.setAttribute('src', url);

    document.body.appendChild(iframe);
    test.add_cleanup(() => iframe.remove());

    return new Promise((resolve) => {
       if (userActivated) {
         test_driver.bless('enable user activation', setSrc);
       } else {
         setSrc();
       }

        iframe.onload = iframe.onerror = resolve;
      });
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite', 'sameOrigin'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameOrigin'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Cross-Site -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'sameSite'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Cross-Site -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['crossSite', 'crossSite'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Cross-Site -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameOrigin'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'sec-fetch-site - Same-Origin -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'sameSite'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - Same-Origin -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameOrigin', 'crossSite'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameOrigin'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - Same-Site -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'sameSite'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'sec-fetch-site - Same-Site -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['sameSite', 'crossSite'], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Same-Site -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, [], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_equals(headers['sec-fetch-mode'], 'navigate');
        });
  }, 'sec-fetch-mode');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, [], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'iframe');
        });
  }, 'sec-fetch-dest');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, [], {mime: 'text/html'}),
        t,
        false
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, [], {mime: 'text/html'}),
        t,
        true
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-user');
          assert_equals(headers['sec-fetch-user'], '?1');
        });
  }, 'sec-fetch-user with user activation');
  </script>
  </body>
</html>
